// blog-fe/src/types/blog.ts
export interface IAuthor{
    objectId: string;
    name: string;
    email: string;
}

export interface IBlog{
    objectId: string;
    title: string;
    category: string;
    content: string;
    thumbnail: string;
    author: IAuthor;
    created: number;
}


export interface BlogInput {
    title: string;
    category: string;
    content: string;
    thumbnail: string;
  }